requirements.txt
tradingview-ta==3.3.0
websocket-client==1.6.4
pandas==2.1.4
colorama==0.4.6  # For colored terminal output
